+++ src/module/gimbal/mod_gimbal.hpp
@@ -66,7 +66,7 @@ class Gimbal {
 
   uint32_t lask_wakeup_;
 
-  uint32_t now_;
+  float now_;
 
   float dt_;
